Contacts(HFWutils)	R Documentation

~~ data name/kind ... ~~

Description

~~ A concise (1-5 lines) description of the dataset. ~~

Usage

data(Contacts)

Format

A data frame with 4 observations on the following 6 variables.

Function

a factor with levels Client acquisition Client service Contact Function

Name

a factor with levels Martin Paul Name Paris Hilton Paul Gerhard

Telephone

a factor with levels 44 1222222222222 44 207 777777 44 26 66666666 Telephone

Fax

a factor with levels 44 1222222222222 44 207 777777 44 26 66666666 Fax

E.Mail

a factor with levels e@company.com E.Mail

Details

example data

Author(s)

Felix Wittmann hfwittmann@gmail.com

Source

~~ reference to a publication or URL from which the data were obtained ~~

References

~~ possibly secondary sources and usages ~~

See Also

df2matrix, logDataframe

Examples


 #  see links

[Package HFWutils version 0.7.2007.09.02 Index]