cleanTexts(HFWutils)	R Documentation

adds spaces removes multiple spaces and trims spaces

Description

vectorised clean up does three things a) adds spaces around special characters ,\.\":!?;()- b) gets rid of multiple spaces c) trims leading and trailing spaces

Usage


cleanTexts(x)

Arguments

x 	 ~~Describe x here~~

Details

~~ If necessary, more details than the description above ~~

Value

~Describe the value returned If it is a LIST, use

comp1 	 Description of 'comp1'

comp2 	 Description of 'comp2'

...

Note

~~further notes~~

Author(s)

Felix Wittmann hfwittmann@gmail.com

References

~put references to the literature/web site here ~

See Also

~~objects to See Also as help, ~~~

Examples


  
  s0 <- c('asd:-asdas','asd?asdas')
  # [1] "asd:-asdas" "asd?asdas" 
  cleanTexts(s0)
  # [1] "asd : - asdas" "asd ? asdas" 


[Package HFWutils version 0.7.2007.09.02 Index]