df2matrix(HFWutils)	R Documentation

data.frame to matrix

Description

converts data.frame to matrix

Usage


df2matrix(x_df)

Arguments

x_df 	 ~~Describe x_df here~~

Details

~~ If necessary, more details than the description above ~~

Value

~Describe the value returned If it is a LIST, use

comp1 	 Description of 'comp1'

comp2 	 Description of 'comp2'

...

Note

~~further notes~~

Author(s)

Felix Wittmann hfwittmann@gmail.com

References

~put references to the literature/web site here ~

See Also

logDataframe, Contacts

Examples



data(Contacts)
Contacts
#    X           Function         Name        Telephone              Fax
# 1 NA           Function         Name        Telephone              Fax
# 2  1            Contact  Martin Paul    44 207 777777    44 207 777777
# 3  2 Client acquisition Paul Gerhard   44 26 66666666   44 26 66666666
# 4  3     Client service Paris Hilton 44 1222222222222 44 1222222222222
#          E.Mail
# 1        E.Mail
# 2 e@company.com
# 3 e@company.com
# 4 e@company.com



[Package HFWutils version 0.7.2007.09.02 Index]