doReplace(HFWutils)	R Documentation

performs replacements in a string

Description

performs replacements in a string using a dataframe set up with initReplaceBy

Usage


doReplace(sqlCommand, ReplaceBy, fixed = TRUE)

Arguments

sqlCommand 	 the target string

ReplaceBy 	 a dataframe set up with initReplaceBy

fixed 	 ~~Describe fixed here~~

Details

~~ If necessary, more details than the description above ~~

Value

~Describe the value returned If it is a LIST, use

comp1 	 Description of 'comp1'

comp2 	 Description of 'comp2'

...

Warning

....

Note

~~further notes~~

~Make other sections like Warning with section{Warning }{....} ~

Author(s)

Felix Wittmann hfwittmann@gmail.com

References

~put references to the literature/web site here ~

See Also

~~objects to See Also as initReplaceBy, ~~~

Examples



      ReplaceBy <- initReplaceBy(2)
      ReplaceBy[1,]  <- c(Replace = "TABLE"          , By = 'USArrests')
      ReplaceBy[2,]  <- c(Replace = "CONDITION"      , By = 'Assault>250')

      sqlCommand <- "SELECT * FROM TABLE WHERE CONDITION"
      doReplace(sqlCommand, ReplaceBy)
      # "SELECT * FROM USArrests WHERE Assault>250"
      

[Package HFWutils version 0.7.2007.09.02 Index]