extractFromList(HFWutils)	R Documentation

extract variables From List

Description

extract data from lists of lists ; these are typically the output of a lapply or sapply

Usage


extractFromList(mIn, names)

Arguments

mIn 	 ~~Describe mIn here~~

names 	 ~~Describe names here~~

Details

~~ If necessary, more details than the description above ~~

Value

~Describe the value returned If it is a LIST, use

comp1 	 Description of 'comp1'

comp2 	 Description of 'comp2'

...

Warning

....

Note

~~further notes~~

~Make other sections like Warning with section{Warning }{....} ~

Author(s)

Felix Wittmann hfwittmann@gmail.com

References

~put references to the literature/web site here ~

See Also

~~objects to See Also as help, ~~~

Examples


    
    c1 <- list(a=1,b=2)
    c2 <- list(a=3,b=4)
    
    d<- list(c1,c2)
    d
    #    [[1]]
    #    [[1]]$a
    #    [1] 1
    #    
    #    [[1]]$b
    #    [1] 2
    #    
    #    
    #    [[2]]
    #    [[2]]$a
    #    [1] 3
    #    
    #    [[2]]$b
    #    [1] 4
    
    extractFromList(d,names=c('a','b'))
    
    #$a
    #     [,1]
    #[1,] 1   
    #[2,] 3   
    #
    #$b
    #     [,1]
    #[1,] 2   
    #[2,] 4  


[Package HFWutils version 0.7.2007.09.02 Index]