initReplaceBy(HFWutils)	R Documentation

sets up dataframe used by doReplace

Description

~~ A concise (1-5 lines) description of what the function does. ~~

Usage


initReplaceBy(n)

Arguments

n 	 ~~Describe n here~~

Details

~~ If necessary, more details than the description above ~~

Value

~Describe the value returned If it is a LIST, use

comp1 	 Description of 'comp1'

comp2 	 Description of 'comp2'

...

Author(s)

Felix Wittmann hfwittmann@gmail.com

References

~put references to the literature/web site here ~

See Also

~~objects to See Also as doReplace, ~~~

Examples




[Package HFWutils version 0.7.2007.09.02 Index]