largeDataframe(HFWutils)	R Documentation

produce dataframe with constant columns

Description

produce dataframe with constant columns. as specified by the input row vector

Usage


largeDataframe(smallDf, Times)

Arguments

smallDf 	 input row vector

Times 	 number of rows

Details

~~ If necessary, more details than the description above ~~

Value

~Describe the value returned If it is a LIST, use

comp1 	 Description of 'comp1'

comp2 	 Description of 'comp2'

...

Note

~~further notes~~

Author(s)

Felix Wittmann hfwittmann@gmail.com

References

~put references to the literature/web site here ~

See Also

~~objects to See Also as help, ~~~

Examples



  
  strategic     <- data.frame(REX = 0.50, DJSTOXX=0.17,SP500=0.17,      SPI=0.08,NIKKEI=0.08)
  
  #  strategic 
  #    REX DJSTOXX SP500  SPI NIKKEI
  #  1 0.5    0.17  0.17 0.08   0.08

  largeDataframe(smallDf=strategic, Times=10)
  #     REX DJSTOXX SP500  SPI NIKKEI
  #  1  0.5    0.17  0.17 0.08   0.08
  #  2  0.5    0.17  0.17 0.08   0.08
  #  3  0.5    0.17  0.17 0.08   0.08
  #  4  0.5    0.17  0.17 0.08   0.08
  #  5  0.5    0.17  0.17 0.08   0.08
  #  6  0.5    0.17  0.17 0.08   0.08
  #  7  0.5    0.17  0.17 0.08   0.08
  #  8  0.5    0.17  0.17 0.08   0.08
  #  9  0.5    0.17  0.17 0.08   0.08
  #  10 0.5    0.17  0.17 0.08   0.08


[Package HFWutils version 0.7.2007.09.02 Index]